android: upgrade gradle plugin runtime to 8.1
Another upgrade recommended by Android Studio, again this is just
upgrading the build tooling, and not yet changing any of the runtime
side dependencies or behaviors.

We switch from jcenter to mavenCentral as advised in the build output,
as jcenter has shutdown.

This includes a mandatory JDK bump from 11 to 20.

Updates #cleanup