android: complete settings implementation
updates ENG-2854

Render the different types of settings.  CorpDNS is working and the about and bug report are stubbed to TBD views.  Error handling not implemented (though we may choose to just swallow API errors here similar to iOS).  Future-proofed to allow us to add any setting we wish to this view.

Styling incomplete.

Corrected the bit mask for the notifier which was off by one bit.

Signed-off-by: Jonathan Nobels <jonathan@tailscale.com>