android/ui: switch lateinint app instance to optional
updates tailscale/corp#12430

The app instance is set in onCreate, but this has a chance to race
with the intent handlers throwing a null pointer exception.

This should ultimately be corrected in the API interface on the caller side, on but
that fix runs deep and still requires care on the part of the caller.  This will,
at least, error out gracefully rather than crashing the application process at the
expense of an action potentially failing.

Signed-off-by: Jonathan Nobels <jonathan@tailscale.com>