mdm: improve handling and returning of not configured policy settings
We should distinguish between unconfigured policy settings and those configured with the default values.
In the first case, the syspolicyHandler should return syspolicy.ErrNoSuchKey, while in the latter case, it should
return the actual setting value, even if that value happens to be the default value such as "user-decides".
This distinction should also be reflected in the "Current MDM settings" view.

In this PR, we update MDMSetting.flow to hold both the value to be used by the app and a flag indicating
whether the policy setting is configured or not. If the policy setting is not configured, the value is the default
value for the setting type. We then use this new flag to decide whether to throw a NoSuchKeyException from
the Android-side syspolicyHandler implementation and how to display the policy setting in the
"Current MDM settings" view.

Additionally, we update MDMSettings.update and MDMSetting.setFrom to avoid calling
app.getEncryptedPrefs (and reading/decrypting the prefs) for every defined MDM setting.

Updates tailscale/tailscale#12687

Signed-off-by: Nick Khyl <nickk@tailscale.com>