libtailscale: use b.netmon and sys.Bus consistently (#717)
fixes tailscale/android#17747

There was a mismatch between the netmon into which we
were injecting network change events, and the netmon that
the eventBus registered with in the userspace engine.

The switch from directly registering callbacks with netmon
to using the eventBus caused a regression where the
injected events would no longer trigger interface rebinds based
on feedback from the Android network monitoring APIs.

Signed-off-by: Jonathan Nobels <jonathan@tailscale.com>